+++ NEWS.md
@@ -1,3 +1,8 @@
+# mlr3 0.8.0-9000
+
+* New methods for `Task`: `$set_col_roles` and `$set_row_roles` as a replacement
+  for the deprecated and less flexible `$set_col_role` and `$set_row_role`.
+
 # mlr3 0.8.0
 
 * `MeasureOOBError` now has set property `minimize` to `TRUE`.